requirements.txt
PyQt6
pyqtgraph
pyserial
